+++ src/Geopilot.Api/StacServices/StacConverter.cs
@@ -73,17 +73,17 @@ public class StacConverter
 
         if (items.Values.Count == 0)
         {
-            var nowTimestamp = DateTime.Now.ToUniversalTime();
+            var nowTimestamp = DateTime.UtcNow;
             var extent = new StacExtent(ToStacSpatialExtent(mandate.SpatialExtent), new StacTemporalExtent(nowTimestamp, nowTimestamp));
             return new StacCollection(collectionId, string.Empty, extent, null, null)
             {
-                Title = mandate.Name,
+                Title = mandate.Name.GetDisplayText(),
             };
         }
         else
         {
             var collection = StacCollection.Create(collectionId, string.Empty, items);
-            collection.Title = mandate.Name;
+            collection.Title = mandate.Name.GetDisplayText();
             return collection;
         }
     }